separate between home and arcades detail page
