+++ lib/views/views.dart
@@ -1,3 +1,5 @@
 export 'home/view.dart';
 export 'login/view.dart';
 export 'register/view.dart';
+
+export 'arcades/home/view.dart';
